Add type=button to input button to prevent triggering of form submission when toggling the select dropdown.
While testing a form with a Select component, we found that when clicking the
toggle for the Select component it was triggering the form submission (as
opposed to clicking the "Submit" button. This was due to the fact that no type
attribute was specified on the button element that triggers the dropdown of
select inputs. This commit adds this type attribute to fix this. It also adds
some additional data-qaids to enable easier testing of Select Inputs.