Allow us to invert colors on dropdown & menu
For the new navigation bar on sofar client we need to have the ability
to invert the colors to a dark theme ffor the dropdown and navigation.

As part of this it was needed to edit the global theme so color is not
applied specifically to all fields to allow inheritence from the parent.
This has been moved to the body selector so it should still get applied
to all elements through inheritence