Implement workaround when using searchable select on mobile devices
The Popper component doesn't behave as expected on mobile browsers when an input
is focused and the soft keyboard is triggered. Additionally, many mobile browsers
treat fixed and absolute positioned elements as static when an input is focused. Because
of all this, the Popper calculations fail, displaying the select options in the wrong position.
This alternative uses a dumber component that requires less calculations and it's a bit less flexible,
in an attempt to give mobile users a better experience when using searchable select fields.